feat(spectral): rework default spectral rules and expose as static "ibm:oas" ruleset
