+++ src/spectral/rulesets/.defaultsForSpectral.yaml
@@ -1,24 +1,94 @@
-extends: [[spectral:oas, off]]
-formats: [oas2, oas3]
-functionsDir: ../functions
+extends: spectral:oas
 rules:
-  no-eval-in-markdown: true
-  no-script-tags-in-markdown: true
-  openapi-tags: true
-  operation-description: true
-  operation-tags: true
+
+  # Original list created from Spectral with:
+  # jq -r '.rules | to_entries | .[] | select(.value.recommended != false) | "  \(.key): off"' src/rulesets/oas/index.json
+
+  # Turn off -- duplicates no_success_response_codes
+  operation-2xx-response: off
+  # Turn off - duplicates non-configurable validation - form-data.js
+  oas2-operation-formData-consume-check: off
+  # Turn off - duplicates non-configurable validation - operation-ids.js
+  operation-operationId-unique: off
+  # Turn off - duplicates non-configurable validation - operations-shared.js
+  operation-parameters: off
+  # Enable with same severity as Spectral
   operation-tag-defined: true
+  # Turn off - duplicates missing_path_parameter
+  path-params: off
+  # Turn off - exclude from ibm:oas
+  info-contact: off
+  # Turn off - exclude from ibm:oas
+  info-description: off
+  # Enable with same severity as Spectral
+  no-eval-in-markdown: true
+  # Enable with same severity as Spectral
+  no-script-tags-in-markdown: true
+  # Enable with same severity as Spectral
+  openapi-tags: true
+  # Enable with same severity as Spectral
+  operation-description: true
+  # Turn off - duplicates operation_id_case_convention, operation_id_naming_convention
+  operation-operationId: off
+  # Turn off - duplicates operation_id_case_convention
+  operation-operationId-valid-in-url: off
+  # Enable with same severity as Spectral
+  operation-tags: true
+  # Turn off - duplicates missing_path_parameter
+  path-declarations-must-exist: off
+  # Enable with same severity as Spectral
   path-keys-no-trailing-slash: true
+  # Turn off - duplicates non-configurable validation - paths.js
+  path-not-include-query: off
+  # Turn off - duplicates $ref_siblings (off by default)
+  no-$ref-siblings: off
+  # Enable with same severity as Spectral
   typed-enum: true
+  # Enable with same severity as Spectral
   oas2-api-host: true
+  # Enable with same severity as Spectral
   oas2-api-schemes: true
+  # Enable with same severity as Spectral
   oas2-host-trailing-slash: true
-  oas2-valid-example: true
+  # Turn off - dupicates non-configurable validation - security-ibm.js
+  oas2-operation-security-defined: off
+  # Turn off
+  oas2-valid-parameter-example: off
+  # Enable with same severity as Spectral
   oas2-valid-definition-example: true
+  # Turn off
+  oas2-valid-response-example: off
+  # Turn off
+  oas2-valid-response-schema-example: off
+  # Enable with same severity as Spectral
   oas2-anyOf: true
+  # Enable with same severity as Spectral
   oas2-oneOf: true
+  # Turn off
+  oas2-schema: off
+  # Turn off - duplicates non-configurable validation in base validator
+  oas2-unused-definition: off
+  # Enable with same severity as Spectral
   oas3-api-servers: true
+  # Enable with same severity as Spectral
   oas3-examples-value-or-externalValue: true
+  # Turn off - dupicates non-configurable validation - security-ibm.js
+  oas3-operation-security-defined: off
+  # Enable with same severity as Spectral
   oas3-server-trailing-slash: true
-  oas3-valid-example: true
+  # Turn off
+  oas3-valid-oas-parameter-example: off
+  # Turn off
+  oas3-valid-oas-header-example: off
+  # Turn off
+  oas3-valid-oas-content-example: off
+  # Turn off
+  oas3-valid-parameter-schema-example: off
+  # Turn off
+  oas3-valid-header-schema-example: off
+  # Enable with same severity as Spectral
   oas3-valid-schema-example: true
+  # Turn off
+  oas3-schema: off
+  # Turn off - duplicates non-configurable validation in base validator
+  oas3-unused-components-schema: off
